+++ HISTORY.rst
@@ -2,6 +2,11 @@
 History
 =======
 
+0.6.0 (2023-06-13)
+------------------
+* Rewrite native part in Rust with support of Py03 and maturin
+* Add async client
+
 0.5.0 (2023-05-19)
 ------------------
 * Move pytogo Go code into portforward